Further separated out platform metric functionality and tests
