+++ .gitignore
@@ -129,3 +129,6 @@ dist
 .yarn/install-state.gz
 .pnp.*
 secrets
+
+# MacOS file for folder view display
+.DS_Store
