+++ lib/platform-stats-fetchers.js
@@ -2,7 +2,7 @@
  * @param {import('pg').Pool} pgPool
  * @param {import('./typings').Filter} filter
  */
-export const fetchDailyNodeMetrics = async (pgPool, filter) => {
+export const fetchDailyStationMetrics = async (pgPool, filter) => {
   const { rows } = await pgPool.query(`
     SELECT day::TEXT, station_id
     FROM daily_node_metrics
