+++ test/test-helpers.js
@@ -1,0 +1,11 @@
+import { AssertionError } from 'node:assert'
+
+export const assertResponseStatus = async (res, status) => {
+  if (res.status !== status) {
+    throw new AssertionError({
+      actual: res.status,
+      expected: status,
+      message: await res.text()
+    })
+  }
+}
